Removed ANN algorithm to separate class
Testing needed
