+++ src/main/java/nl/tudelft/bep/deeplearning/clustering/Mapping.java
@@ -8,9 +8,7 @@ import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Scanner;
 
@@ -22,7 +20,6 @@ import nl.tudelft.bep.deeplearning.clustering.exception.MinimumNotPossibleExcept
 
 public final class Mapping {
 
-	private static final int NEIGHBORS_STRIP = 8;
 	private static final int BATCH_SIZE = 32;
 	private static final double TRAINING_SET = 0.7;
 	private static int outputNum;
@@ -212,7 +209,7 @@ public final class Mapping {
 	 */
 	public static Cluster[] createClusters(final Cluster[] layer1) {
 		// Run nearest neighbor algorithm.
-		List<Neighbors> neighbors = neighbour(layer1);
+		List<Neighbors> neighbors = NearestNeighbors.neighbour(layer1);
 		// Temporary ArrayList for the newly created higher level clusters.
 		List<Cluster> layer2 = new ArrayList<Cluster>();
 		Collections.sort(neighbors);
@@ -239,24 +236,7 @@ public final class Mapping {
 		return res;
 	}
 
-	/**
-	 * Helper function which finds the cluster in an array based on its ID.
-	 * 
-	 * @param list
-	 *            list of clusters where we would like to find the cluster with
-	 *            ID id
-	 * @param id
-	 *            the id we would like to find
-	 * @return the index in the array
-	 */
-	public static int getIndexFromID(final Cluster[] list, final int id) {
-		for (int i = 0; i < list.length; i++) {
-			if (list[i].getID() == id) {
-				return i;
-			}
-		}
-		return -1;
-	}
+
 
 	/**
 	 * Helper function which finds the cluster in an list of Neighbors.class.
@@ -324,105 +304,5 @@ public final class Mapping {
 		return matrix;
 	}
 
-	/**
-	 * Returns the nearest neighbor for each cluster or point as an array.
-	 * 
-	 * @param clusters
-	 *            The clusters we would like to know the nearest neighbors of.
-	 * @return Array of nearest neighbors
-	 */
-	public static List<Neighbors> neighbour(final Cluster[] clusters) {
-		int[] result = new int[clusters.length];
-		// Distances is used to keep track of the closest distance found for
-		// each point.
-		Double[] distances = new Double[clusters.length];
-		for (int i = 0; i < clusters.length; i++) {
-			distances[i] = Double.MAX_VALUE;
-		}
-		closestClusters(clusters.clone(), result, distances);
-		List<Neighbors> neighbors = new ArrayList<Neighbors>();
-		for (int i = 0; i < clusters.length; i++) {
-			int index = getIndexFromID(clusters, result[i]);
-			neighbors.add(new Neighbors(clusters[i], clusters[index], distances[i]));
-		}
-		return neighbors;
-	}
 
-	/**
-	 * Recursive function for finding the nearest neighbor for each
-	 * cluster/point.
-	 * 
-	 * @param clusters
-	 *            Array containing the cluster we would like to know the nearest
-	 *            neighbor of.
-	 * @param result
-	 *            Array where we would like to keep track of the nearest
-	 *            neighbor found yet.
-	 * @param distances
-	 *            Array for keeping track of the distances of the result array.
-	 */
-	public static void closestClusters(final Cluster[] clusters, final int[] result, final Double[] distances) {
-
-		int n = clusters.length;
-
-		if (n == 1) {
-			return;
-		}
-
-		java.util.Arrays.sort(clusters);
-
-		int mid = clusters.length / 2;
-		Cluster median = clusters[mid];
-		Cluster[] part1 = new Cluster[mid];
-		Cluster[] part2 = new Cluster[n - mid];
-
-		part1 = Arrays.copyOfRange(clusters, 0, mid);
-		part2 = Arrays.copyOfRange(clusters, mid, n);
-
-		closestClusters(part1, result, distances);
-		closestClusters(part2, result, distances);
-
-		ArrayList<Cluster> strip = new ArrayList<Cluster>();
-
-		double max = 0;
-		for (int i = 0; i < n; i++) {
-			double distance = distances[clusters[i].getID()];
-			if (distance > max) {
-				max = distance;
-			}
-		}
-
-		for (int i = 0; i < n; i++) {
-			if (Math.abs(median.getX() - clusters[i].getX()) < max) {
-				strip.add(clusters[i]);
-			}
-		}
-
-		Collections.sort(strip, new Comparator<Cluster>() {
-			public int compare(final Cluster p, final Cluster p2) {
-				return Double.compare(p.getY(), p2.getY());
-			}
-		});
-
-		for (int i = 0; i < strip.size(); i++) {
-			int k = strip.size() - i - 1;
-			if (k > NEIGHBORS_STRIP) {
-				k = NEIGHBORS_STRIP + 1 + i;
-			} else {
-				k = strip.size();
-			}
-			for (int j = i + 1; j < k; j++) {
-				double tempDistance = Math.sqrt(Math.pow(strip.get(j).getX() - strip.get(i).getX(), 2)
-						+ Math.pow(strip.get(j).getY() - strip.get(i).getY(), 2));
-				if (distances[strip.get(i).getID()] > tempDistance) {
-					result[strip.get(i).getID()] = strip.get(j).getID();
-					distances[strip.get(i).getID()] = tempDistance;
-				}
-				if (distances[strip.get(j).getID()] > tempDistance) {
-					result[strip.get(j).getID()] = strip.get(i).getID();
-					distances[strip.get(j).getID()] = tempDistance;
-				}
-			}
-		}
-	}
 }
